+++ pages/_app.js
@@ -1,13 +1,13 @@
+import Head from 'next/head'
+import Layout from '../components/layout/Layout'
+import favicon from '../public/favicon.ico'
+import { createTheme, ThemeProvider, responsiveFontSizes } from '@mui/material/styles'
 import '../styles/globals.css'
 import '../components/header/header.scss'
 import '../styles/pagesStyles/home.scss'
 import '../components/footer/footer.scss'
 import '../styles/pagesStyles/contactUs.scss'
 import '../styles/pagesStyles/careers.scss'
-import { createTheme, ThemeProvider, responsiveFontSizes } from '@mui/material/styles'
-import Head from 'next/head'
-import Layout from '../components/layout/Layout'
-import favicon from '../public/favicon.ico'
 
 let theme = createTheme({
   breakpoints: {
