+++ pages/index.js
@@ -3,7 +3,6 @@ import Image from 'next/image'
 import Header from '../components/header/Header'
 import homePageBg from "../public/images/homePageBg.png";
 import ArrowRightAltIcon from "@mui/icons-material/ArrowRightAlt";
-import Script from 'next/script';
 import { Box, Container, Typography, Stack, Divider, Grid, Button } from '@mui/material';
 import { homeData, services, servicesHeading } from '../constants/constants';
 import { useRouter } from 'next/router';
@@ -20,17 +19,6 @@ export default function Home() {
         <title>{title}</title>
         <meta name='prixite' content='Prixite' />
         <link rel='icon' href='/favicon.ico' />
-        <Script strategy="lazyOnload" src={`https://www.googletagmanager.com/gtag/js?id='${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}'`}></Script>
-        <Script strategy="lazyOnload">
-                {`
-                    window.dataLayer = window.dataLayer || [];
-                    function gtag(){dataLayer.push(arguments);}
-                    gtag('js', new Date());
-                    gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}', {
-                    page_path: window.location.pathname,
-                    });
-                `}
-        </Script>
       </Head>
 
       <Header bgImg={homePageBg}>
